refactor: modularize DataExportTab component for better maintainability
- Created useDataExport composable for export logic
- Extracted JSON export functionality to utils/exportJSON.js
- Extracted PDF export functionality to utils/exportPDF.js
- Reduced DataExportTab.vue from 400+ lines to ~130 lines
- Improved code organization and separation of concerns
- Enhanced maintainability and testability

Benefits:
- Export logic can now be reused across different components
- Easier to unit test export functions independently
- Cleaner component focused solely on UI presentation
- Better follows single responsibility principle
